A few doc updates
